requirements.txt
PyQt6>=6.4.0
pyautogui>=0.9.54
Pillow>=10.0.0
keyboard>=0.4.1
pyinstaller>=6.0.0 